pyproject.toml
[project]
name = "hng-stage-0"
version = "0.1.0"
description = "The first Task for HNG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudpathlib==0.16.0",
    "fastapi>=0.119.0",
    "google-auth>=2.41.1",
    "google-genai>=1.46.0",
    "numpy==1.26.4",
    "pytest>=8.4.2",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "smart-open==6.4.0",
    "sqlmodel>=0.0.27",
    "thinc==8.2.2",
    "typer==0.9.4",
    "uvicorn>=0.37.0",
    "weasel==0.3.4",
]
